requirements.txt
requests
beautifulsoup4
python-telegram-bot
soundfile
sounddevice